+++ aioxmpp/muc/__init__.py
@@ -71,6 +71,9 @@ Forms
 .. autoclass:: ConfigurationForm
    :members:
 
+.. autoclass:: InfoForm
+    :members:
+
 .. currentmodule:: aioxmpp.muc.xso
 
 XSOs
@@ -122,6 +125,7 @@ Owner namespace
 from .service import MUCClient, Occupant, Room, LeaveMode, RoomState  # NOQA
 from . import xso  # NOQA
 from .xso import (  # NOQA
-    ConfigurationForm
+    ConfigurationForm,
+    InfoForm,
 )
 Service = MUCClient  # NOQA